[NO-TASK] More cleanup
