+++ src/subdomains/generic/kyc/kyc.module.ts
@@ -1,6 +1,5 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { BlockchainModule } from 'src/integration/blockchain/blockchain.module';
 import { SharedModule } from 'src/shared/shared.module';
 import { BuyCryptoModule } from 'src/subdomains/core/buy-crypto/buy-crypto.module';
 import { SellCryptoModule } from 'src/subdomains/core/sell-crypto/sell-crypto.module';
@@ -54,7 +53,6 @@ import { TfaService } from './services/tfa.service';
       KycFile,
     ]),
     SharedModule,
-    BlockchainModule,
     NotificationModule,
     forwardRef(() => UserModule),
     forwardRef(() => BuyCryptoModule),
