+++ src/subdomains/supporting/realunit/realunit.service.ts
@@ -2,17 +2,23 @@ import { BadRequestException, Injectable, NotFoundException } from '@nestjs/comm
 import { verifyTypedData } from 'ethers/lib/utils';
 import { request } from 'graphql-request';
 import { GetConfig } from 'src/config/config';
+import {
+  AllowlistStatusDto,
+  BrokerbotBuyPriceDto,
+  BrokerbotInfoDto,
+  BrokerbotPriceDto,
+  BrokerbotSharesDto,
+} from 'src/integration/blockchain/realunit/dto/realunit-broker.dto';
 import { RealUnitBlockchainService } from 'src/integration/blockchain/realunit/realunit-blockchain.service';
 import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
 import { Asset, AssetType } from 'src/shared/models/asset/asset.entity';
 import { AssetService } from 'src/shared/models/asset/asset.service';
 import { AsyncCache, CacheItemResetPeriod } from 'src/shared/utils/async-cache';
 import { Util } from 'src/shared/utils/util';
-import { RealUnitRegistrationDto } from 'src/subdomains/generic/kyc/dto/input/realunit-registration.dto';
 import { KycStepName } from 'src/subdomains/generic/kyc/enums/kyc-step-name.enum';
 import { ReviewStatus } from 'src/subdomains/generic/kyc/enums/review-status.enum';
 import { KycService } from 'src/subdomains/generic/kyc/services/kyc.service';
-import { UserDataService } from 'src/subdomains/generic/user/models/user-data/user-data.service';
+import { UserService } from 'src/subdomains/generic/user/models/user/user.service';
 import { AssetPricesService } from '../pricing/services/asset-prices.service';
 import { PriceCurrency, PriceValidity, PricingService } from '../pricing/services/pricing.service';
 import {
@@ -22,15 +28,11 @@ import {
   TokenInfoClientResponse,
 } from './dto/client.dto';
 import { RealUnitDtoMapper } from './dto/realunit-dto.mapper';
+import { RealUnitRegistrationDto } from './dto/realunit-registration.dto';
 import {
   AccountHistoryDto,
   AccountSummaryDto,
-  AllowlistStatusDto,
   BankDetailsDto,
-  BrokerbotBuyPriceDto,
-  BrokerbotInfoDto,
-  BrokerbotPriceDto,
-  BrokerbotSharesDto,
   HistoricalPriceDto,
   HoldersDto,
   TimeFrame,
@@ -51,7 +53,7 @@ export class RealUnitService {
     private readonly pricingService: PricingService,
     private readonly assetService: AssetService,
     private readonly blockchainService: RealUnitBlockchainService,
-    private readonly userDataService: UserDataService,
+    private readonly userService: UserService,
     private readonly kycService: KycService,
   ) {
     this.ponderUrl = GetConfig().blockchain.realunit.graphUrl;
@@ -165,22 +167,19 @@ export class RealUnitService {
   // --- Registration Methods ---
 
   async register(userDataId: number, dto: RealUnitRegistrationDto): Promise<void> {
-    const userData = await this.userDataService.getUserData(userDataId, { users: true, kycSteps: true });
-    if (!userData) throw new NotFoundException('User not found');
+    const userData = await this.userService
+      .getUserByAddress(dto.walletAddress, { userData: { kycSteps: true } })
+      .then((u) => u?.userData);
 
-    const walletBelongsToUser = userData.users.some((u) => u.address.toLowerCase() === dto.walletAddress.toLowerCase());
-    if (!walletBelongsToUser) throw new BadRequestException('Wallet address does not belong to user');
+    if (!userData) throw new NotFoundException('User not found');
+    if (userData.id !== userDataId) throw new BadRequestException('Wallet address does not belong to user');
 
     // verify EIP-712 signature
     const isValidSignature = this.verifyRealUnitRegistrationSignature(dto);
     if (!isValidSignature) throw new BadRequestException('Invalid signature');
 
     // check for existing registration
-    const existingStep = userData.kycSteps?.find(
-      (s) =>
-        s.name === KycStepName.REALUNIT_REGISTRATION &&
-        ![ReviewStatus.FAILED, ReviewStatus.CANCELED].includes(s.status),
-    );
+    const existingStep = userData.getNonFailedStepWith(KycStepName.REALUNIT_REGISTRATION);
     if (existingStep) throw new BadRequestException('RealUnit registration already exists');
 
     // store data
@@ -219,6 +218,6 @@ export class RealUnitService {
     const signatureToUse = data.signature.startsWith('0x') ? data.signature : `0x${data.signature}`;
     const recoveredAddress = verifyTypedData(domain, types, data, signatureToUse);
 
-    return recoveredAddress.toLowerCase() === data.walletAddress.toLowerCase();
+    return Util.equalsIgnoreCase(recoveredAddress, data.walletAddress);
   }
 }
